requirements.txt
google-generativeai==0.7.2
pandas==2.2.3
pyyaml==6.0.2
python-dotenv==1.0.1
langfuse==3.8.5
